Enum ethvm::OpCode

pub enum OpCode {
    STOP,
    ADD,
    MUL,
    SUB,
    DIV,
    SDIV,
    MOD,
    SMOD,
    ADDMOD,
    MULMOD,
    EXP,
    SIGNEXTEND,
    LT,
    GT,
    SLT,
    SGT,
    EQ,
    ISZERO,
    AND,
    OR,
    XOR,
    NOT,
    BYTE,
    SHA3,
    ADDRESS,
    BALANCE,
    ORIGIN,
    CALLER,
    CALLVALUE,
    CALLDATALOAD,
    CALLDATASIZE,
    CALLDATACOPY,
    CODESIZE,
    CODECOPY,
    GASPRICE,
    EXTCODESIZE,
    EXTCODECOPY,
    RETURNDATASIZE,
    RETURNDATACOPY,
    BLOCKHASH,
    COINBASE,
    TIMESTAMP,
    NUMBER,
    DIFFICULTY,
    GASLIMIT,
    POP,
    MLOAD,
    MSTORE,
    MSTORE8,
    SLOAD,
    SSTORE,
    JUMP,
    JUMPI,
    PC,
    MSIZE,
    GAS,
    JUMPDEST,
    PUSH1([u8; 1]),
    PUSH2([u8; 2]),
    PUSH3([u8; 3]),
    PUSH4([u8; 4]),
    PUSH5([u8; 5]),
    PUSH6([u8; 6]),
    PUSH7([u8; 7]),
    PUSH8([u8; 8]),
    PUSH9([u8; 9]),
    PUSH10([u8; 10]),
    PUSH11([u8; 11]),
    PUSH12([u8; 12]),
    PUSH13([u8; 13]),
    PUSH14([u8; 14]),
    PUSH15([u8; 15]),
    PUSH16([u8; 16]),
    PUSH17([u8; 17]),
    PUSH18([u8; 18]),
    PUSH19([u8; 19]),
    PUSH20([u8; 20]),
    PUSH21([u8; 21]),
    PUSH22([u8; 22]),
    PUSH23([u8; 23]),
    PUSH24([u8; 24]),
    PUSH25([u8; 25]),
    PUSH26([u8; 26]),
    PUSH27([u8; 27]),
    PUSH28([u8; 28]),
    PUSH29([u8; 29]),
    PUSH30([u8; 30]),
    PUSH31([u8; 31]),
    PUSH32([u8; 32]),
    DUP1,
    DUP2,
    DUP3,
    DUP4,
    DUP5,
    DUP6,
    DUP7,
    DUP8,
    DUP9,
    DUP10,
    DUP11,
    DUP12,
    DUP13,
    DUP14,
    DUP15,
    DUP16,
    SWAP1,
    SWAP2,
    SWAP3,
    SWAP4,
    SWAP5,
    SWAP6,
    SWAP7,
    SWAP8,
    SWAP9,
    SWAP10,
    SWAP11,
    SWAP12,
    SWAP13,
    SWAP14,
    SWAP15,
    SWAP16,
    LOG0,
    LOG1,
    LOG2,
    LOG3,
    LOG4,
    CREATE,
    CALL,
    CALLCODE,
    RETURN,
    DELEGATECALL,
    STATICCALL,
    REVERT,
    INVALID,
    SELFDESTRUCT,
    BAD(u8),
}

Include an instruction and its immediate values if exist.

More details can be found in the chapter Appendix H. Virtual Machine Specification of Ethereum Yellow Paper.

Defined by the proc-macro instruction_set.

Variants
STOPADDMULSUBDIVSDIVMODSMODADDMODMULMODEXPSIGNEXTENDLTGTSLTSGTEQISZEROANDORXORNOTBYTESHA3ADDRESSBALANCEORIGINCALLERCALLVALUECALLDATALOADCALLDATASIZECALLDATACOPYCODESIZECODECOPYGASPRICEEXTCODESIZEEXTCODECOPYRETURNDATASIZERETURNDATACOPYBLOCKHASHCOINBASETIMESTAMPNUMBERDIFFICULTYGASLIMITPOPMLOADMSTOREMSTORE8SLOADSSTOREJUMPJUMPIPCMSIZEGASJUMPDESTPUSH1([u8; 1])PUSH2([u8; 2])PUSH3([u8; 3])PUSH4([u8; 4])PUSH5([u8; 5])PUSH6([u8; 6])PUSH7([u8; 7])PUSH8([u8; 8])PUSH9([u8; 9])PUSH10([u8; 10])PUSH11([u8; 11])PUSH12([u8; 12])PUSH13([u8; 13])PUSH14([u8; 14])PUSH15([u8; 15])PUSH16([u8; 16])PUSH17([u8; 17])PUSH18([u8; 18])PUSH19([u8; 19])PUSH20([u8; 20])PUSH21([u8; 21])PUSH22([u8; 22])PUSH23([u8; 23])PUSH24([u8; 24])PUSH25([u8; 25])PUSH26([u8; 26])PUSH27([u8; 27])PUSH28([u8; 28])PUSH29([u8; 29])PUSH30([u8; 30])PUSH31([u8; 31])PUSH32([u8; 32])DUP1DUP2DUP3DUP4DUP5DUP6DUP7DUP8DUP9DUP10DUP11DUP12DUP13DUP14DUP15DUP16SWAP1SWAP2SWAP3SWAP4SWAP5SWAP6SWAP7SWAP8SWAP9SWAP10SWAP11SWAP12SWAP13SWAP14SWAP15SWAP16LOG0LOG1LOG2LOG3LOG4CREATECALLCALLCODERETURNDELEGATECALLSTATICCALLREVERTINVALIDSELFDESTRUCTBAD(u8)